Isolate SECRET_KEY value in `.env`.
Instead of defining the value of the SECRET_KEY in setup/settings.py,
the `.env` file is used instead, so the information is not exposed in
the repository.

Note that the `.env` file should never be included in the git
repository, but defined in the locally in the server

The `requirements.txt` is also updated o include the python-dotevn
package, used to load the content of `.env` as environment variables.